app done

+++ app/src/main/java/uz/texnopos/contactsandroid4/ui/all/AllContactsFragment.kt
@@ -1,8 +1,11 @@
 package uz.texnopos.contactsandroid4.ui.all
 
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.snackbar.Snackbar
 import uz.texnopos.contactsandroid4.R
 import uz.texnopos.contactsandroid4.data.ContactDao
@@ -26,13 +29,51 @@ class AllContactsFragment : Fragment(R.layout.fragment_contacts_all) {
         binding.apply {
             recyclerView.adapter = adapter
 
-            adapter.setOnDeleteClickListener { contact ->
-                contactDao.deleteContact(contact)
-                Snackbar.make(
-                    fabAddContact,
-                    "Contact deleted successfully!",
-                    Snackbar.LENGTH_SHORT
-                ).show()
+            adapter.setOnDeleteClickListener { contact, position ->
+                MaterialAlertDialogBuilder(requireContext())
+                    .setTitle("Delete contact")
+                    .setMessage("Are you sure to delete ${contact.name} from your contacts list?")
+                    .setPositiveButton("Yes") { dialog, _ ->
+                        contactDao.deleteContact(contact)
+                        adapter.removeAtPosition(position)
+                        Snackbar.make(
+                            fabAddContact,
+                            "Contact deleted successfully!",
+                            Snackbar.LENGTH_SHORT
+                        ).show()
+                        dialog.dismiss()
+                    }
+                    .setNegativeButton("Cancel") { dialog, _ ->
+                        dialog.dismiss()
+                    }
+                    .show()
+            }
+
+            adapter.setOnFavoriteClickListener { contact, position ->
+                val newContact = contact.copy(isFavorite = 1 - contact.isFavorite)
+                contactDao.updateContact(newContact)
+                adapter.updateContact(position)
+            }
+
+            adapter.setOnPhoneClickListener { phoneNumber ->
+                val uri = "tel:$phoneNumber"
+                val intent = Intent(Intent.ACTION_DIAL, Uri.parse(uri))
+                startActivity(intent)
+            }
+
+            adapter.setOnItemClickListener { contact, position ->
+                val bundle = Bundle()
+
+                bundle.putInt("id", contact.id)
+                bundle.putString("name", contact.name)
+                bundle.putString("phone", contact.phone)
+                bundle.putString("image", contact.image)
+                bundle.putInt("isFavorite", contact.isFavorite)
+
+                requireActivity().supportFragmentManager.beginTransaction()
+                    .replace(R.id.fragment_container, EditContactFragment::class.java, bundle)
+                    .addToBackStack(EditContactFragment::class.java.simpleName)
+                    .commit()
             }
 
             adapter.contacts = contactDao.getAllContacts().toMutableList()
